+++ swap/src/cli/api/request.rs
@@ -9,8 +9,9 @@ use crate::cli::{list_sellers as list_sellers_impl, EventLoop, SellerStatus};
 use crate::common::{get_logs, redact};
 use crate::libp2p_ext::MultiAddrExt;
 use crate::monero::wallet_rpc::MoneroDaemon;
-use crate::network::quote::BidQuote;
 use crate::network::rendezvous::XmrBtcNamespace;
+use crate::monero::MoneroAddressPool;
+use crate::network::quote::{BidQuote, ZeroQuoteReceived};
 use crate::network::swarm;
 use crate::protocol::bob::{BobState, Swap};
 use crate::protocol::{bob, Database, State};
@@ -66,8 +67,7 @@ pub struct BuyXmrArgs {
     pub sellers: Vec<Multiaddr>,
     #[typeshare(serialized_as = "Option<string>")]
     pub bitcoin_change_address: Option<bitcoin::Address<NetworkUnchecked>>,
-    #[typeshare(serialized_as = "string")]
-    pub monero_receive_address: monero::Address,
+    pub monero_receive_pool: MoneroAddressPool,
 }
 
 #[typeshare]
@@ -238,6 +238,7 @@ pub struct GetSwapInfoResponse {
     pub cancel_timelock: CancelTimelock,
     pub punish_timelock: PunishTimelock,
     pub timelock: Option<ExpiredTimelocks>,
+    pub monero_receive_pool: MoneroAddressPool,
 }
 
 impl Request for GetSwapInfoArgs {
@@ -575,6 +576,8 @@ pub async fn get_swap_info(
 
     let timelock = swap_state.expired_timelocks(bitcoin_wallet.clone()).await?;
 
+    let monero_receive_pool = context.db.get_monero_address_pool(args.swap_id).await?;
+
     Ok(GetSwapInfoResponse {
         swap_id: args.swap_id,
         seller: AliceAddress {
@@ -594,6 +597,7 @@ pub async fn get_swap_info(
         cancel_timelock,
         punish_timelock,
         timelock,
+        monero_receive_pool,
     })
 }
 
@@ -609,8 +613,11 @@ pub async fn buy_xmr(
         rendezvous_points,
         sellers,
         bitcoin_change_address,
-        monero_receive_address,
+        monero_receive_pool,
     } = buy_xmr;
+
+    monero_receive_pool.assert_network(context.config.env_config.monero_network)?;
+    monero_receive_pool.assert_sum_to_one()?;
 
     let bitcoin_wallet = Arc::clone(
         context
@@ -764,7 +771,7 @@ pub async fn buy_xmr(
 
     context
         .db
-        .insert_monero_address(swap_id, monero_receive_address)
+        .insert_monero_address_pool(swap_id, monero_receive_pool.clone())
         .await?;
 
     tracing::debug!(peer_id = %swarm.local_peer_id(), "Network layer initialized");
@@ -813,7 +820,7 @@ pub async fn buy_xmr(
                     monero_wallet,
                     env_config,
                     event_loop_handle,
-                    monero_receive_address,
+                    monero_receive_pool.clone(),
                     bitcoin_change_address_for_spawn,
                     tx_lock_amount,
                     tx_lock_fee
@@ -888,7 +895,7 @@ pub async fn resume_swap(
     let (event_loop, event_loop_handle) =
         EventLoop::new(swap_id, swarm, seller_peer_id, context.db.clone())?;
 
-    let monero_receive_address = context.db.get_monero_address(swap_id).await?;
+    let monero_receive_pool = context.db.get_monero_address_pool(swap_id).await?;
 
     let swap = Swap::from_db(
         Arc::clone(&context.db),
@@ -906,7 +913,7 @@ pub async fn resume_swap(
             .clone(),
         context.config.env_config,
         event_loop_handle,
-        monero_receive_address,
+        monero_receive_pool,
     )
     .await?
     .with_event_emitter(context.tauri_handle.clone());